// src/lib/client.types.ts
  export type DirectusFiles = {
    charset?: string | null;
    description?: string | null;
    duration?: number | null;
    embed?: string | null;
    filename_disk?: string | null;
    filename_download: string;
    filesize?: number | null;
    folder?: string | DirectusFolders | null;
    height?: number | null;
    id: string;
    location?: string | null;
    metadata?: unknown | null;
    modified_by?: string | DirectusUsers | null;
    modified_on: string;
    storage: string;
    tags?: unknown | null;
    title?: string | null;
    type?: string | null;
    uploaded_by?: string | DirectusUsers | null;
    uploaded_on: string;
    width?: number | null;
  };
// ...
  export type DirectusRoles = {
    admin_access: boolean;
    app_access: boolean;
    description?: string | null;
    enforce_tfa: boolean;
    icon: string;
    id: string;
    ip_access?: unknown | null;
    name: string;
    users?: any[] | DirectusUsers[] | null;
  };
// ...
  export type DirectusUsers = {
    auth_data?: unknown | null;
    avatar?: string | DirectusFiles | null;
    description?: string | null;
    email?: string | null;
    email_notifications?: boolean | null;
    external_identifier?: string | null;
    first_name?: string | null;
    id: string;
    language?: string | null;
    last_access?: string | null;
    last_name?: string | null;
    last_page?: string | null;
    location?: string | null;
    password?: string | null;
    provider: string;
    role?: string | DirectusRoles | null;
    status: string;
    tags?: unknown | null;
    tfa_secret?: string | null;
    theme?: string | null;
    title?: string | null;
    token?: string | null;
  };
  
  export type DirectusWebhooks = {
    actions: unknown;
    collections: unknown;
    data: boolean;
    headers?: unknown | null;
    id: number;
    method: string;
    name: string;
    status: string;
    url: string;
  };
  
  export type Languages = {
    code?: string | null;
    id: number;
    name?: string | null;
  };
  
  export type MenuMondayBreakfast = {
    id: number;
    produits_id?: string | Produits | null;
    restaurants_id?: number | Restaurants | null;
  };
  
  export type MenuMondayDinner = {
    id: number;
    produits_id?: string | Produits | null;
    restaurants_id?: number | Restaurants | null;
  };
  
  export type MenuMondayLunch = {
    id: number;
    produits_id?: string | Produits | null;
    restaurants_id?: number | Restaurants | null;
  };
  
  export type MenuTuesdayBreakfast = {
    id: number;
    produits_id?: string | Produits | null;
    restaurants_id?: number | Restaurants | null;
  };
  
  export type Produits = {
    calory?: string | null;
    category?: number | Category | null;
    date_created?: string | null;
    date_updated?: string | null;
    energy?: string | null;
    i18n_name?: any[] | ProduitsTranslations[] | null;
    id: string;
    name?: string | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
  };
// ...
  export type Restaurants = {
    close_hour?: string | null;
    code_affluence?: number | null;
    date_created?: string | null;
    date_updated?: string | null;
    id: number;
    isFree?: boolean | null;
    monday: string;
    monday_breakfast?: any[] | MenuMondayBreakfast[] | null;
    monday_dinner?: any[] | MenuMondayDinner[] | null;
    monday_lunch?: any[] | MenuMondayLunch[] | null;
    name?: string | null;
    open_hour?: string | null;
    owner?: string | DirectusUsers | null;
    position?: string | null;
    "tabs-kqfgwk": string;
    tuesday: string;
    tuesday_breakfast?: any[] | MenuTuesdayBreakfast[] | null;
    user_created?: string | DirectusUsers | null;
    user_updated?: string | DirectusUsers | null;
  };